requirements.txt
aiohttp>=3.12.15
python-dotenv>=1.0.1
pydantic>=2.11.7
kit-api @ git+https://github.com/lok101/kit-vending-api.git